+++ databarn/__init__.py
@@ -11,14 +11,17 @@ from .cob import Cob
 from .grain import create_grain_class as Grain
 from .barn import Barn
 from .funcs import dict_to_cob, json_to_cob
-from .decorators import one_to_many_grain, one_to_one_grain, post_init
+from .decorators import (
+    one_to_many_grain, one_to_one_grain,
+    post_init, before_assign, after_assign)
 from .exceptions import (
-    DataBarnViolationError, DataBarnSyntaxError,
-    CobConsistencyError, StaticModelViolationError,
-    CobConstraintViolationError, GrainTypeMismatchError,
-    GrainLabelError, BarnConstraintViolationError)
+    DataBarnViolationError,  ValidationError,
+    DataBarnSyntaxError, CobConsistencyError,
+    StaticModelViolationError, CobConstraintViolationError,
+    GrainTypeMismatchError, GrainLabelError,
+    BarnConstraintViolationError)
 
-__version__ = "1.9.2"
+__version__ = "1.10.0"
 
 __all__ = [
     "Barn",
@@ -30,6 +33,7 @@ __all__ = [
     "CobConstraintViolationError",
     "DataBarnSyntaxError",
     "DataBarnViolationError",
+    "ValidationError",
     "GrainLabelError",
     "GrainTypeMismatchError",
     "StaticModelViolationError",
@@ -38,4 +42,6 @@ __all__ = [
     "post_init",
     "one_to_many_grain",
     "one_to_one_grain",
+    "before_assign",
+    "after_assign",
 ]